chore: Extract TenTenOneApp from main.dart
