+++ mobile/lib/common/ten_ten_one_app.dart
@@ -1,0 +1,278 @@
+import 'package:flutter_native_splash/flutter_native_splash.dart';
+import 'package:flutter_rust_bridge/flutter_rust_bridge.dart';
+import 'package:get_10101/features/trade/domain/position.dart';
+import 'package:get_10101/bridge_generated/bridge_definitions.dart' as bridge;
+import 'package:f_logs/f_logs.dart';
+import 'package:get_10101/common/application/channel_info_service.dart';
+import 'package:get_10101/common/application/event_service.dart';
+import 'package:get_10101/common/domain/init_service.dart';
+import 'package:get_10101/features/trade/candlestick_change_notifier.dart';
+import 'package:flutter/material.dart';
+import 'package:get_10101/features/trade/order_change_notifier.dart';
+import 'package:get_10101/features/trade/position_change_notifier.dart';
+import 'package:get_10101/features/trade/submit_order_change_notifier.dart';
+import 'package:get_10101/features/trade/trade_value_change_notifier.dart';
+import 'package:get_10101/features/wallet/wallet_change_notifier.dart';
+import 'package:get_10101/common/color.dart';
+import 'package:get_10101/common/scaffold_with_nav_bar.dart';
+import 'package:get_10101/features/trade/settings_screen.dart';
+import 'package:get_10101/features/trade/trade_theme.dart';
+import 'package:get_10101/features/wallet/create_invoice_screen.dart';
+import 'package:get_10101/features/wallet/domain/share_invoice.dart';
+import 'package:get_10101/features/wallet/seed_screen.dart';
+import 'package:get_10101/features/wallet/share_invoice_screen.dart';
+import 'package:get_10101/features/wallet/scanner_screen.dart';
+import 'package:get_10101/features/wallet/send_screen.dart';
+import 'package:get_10101/features/wallet/settings_screen.dart';
+import 'package:get_10101/features/trade/trade_screen.dart';
+import 'package:get_10101/features/wallet/wallet_screen.dart';
+import 'package:get_10101/features/wallet/wallet_theme.dart';
+import 'package:get_10101/util/preferences.dart';
+import 'package:go_router/go_router.dart';
+import 'package:get_10101/features/welcome/welcome_screen.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+import 'package:provider/provider.dart';
+import 'package:get_10101/features/trade/domain/order.dart';
+import 'package:get_10101/features/trade/domain/price.dart';
+import 'package:get_10101/features/wallet/domain/wallet_info.dart';
+
+final GlobalKey<NavigatorState> _rootNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'root');
+final GlobalKey<NavigatorState> _shellNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'shell');
+
+class TenTenOneApp extends StatefulWidget {
+  const TenTenOneApp({Key? key}) : super(key: key);
+
+  @override
+  State<TenTenOneApp> createState() => _TenTenOneAppState();
+}
+
+class _TenTenOneAppState extends State<TenTenOneApp> {
+  final GoRouter _router = GoRouter(
+      navigatorKey: _rootNavigatorKey,
+      initialLocation: WalletScreen.route,
+      routes: <RouteBase>[
+        ShellRoute(
+          navigatorKey: _shellNavigatorKey,
+          builder: (BuildContext context, GoRouterState state, Widget child) {
+            return ScaffoldWithNavBar(
+              child: child,
+            );
+          },
+          routes: <RouteBase>[
+            GoRoute(
+              path: WalletScreen.route,
+              builder: (BuildContext context, GoRouterState state) {
+                return const WalletScreen();
+              },
+              routes: <RouteBase>[
+                GoRoute(
+                  path: SendScreen.subRouteName,
+                  // Use root navigator so the screen overlays the application shell
+                  parentNavigatorKey: _rootNavigatorKey,
+                  builder: (BuildContext context, GoRouterState state) {
+                    return const SendScreen();
+                  },
+                ),
+                GoRoute(
+                  path: SeedScreen.subRouteName,
+                  // Use root navigator so the screen overlays the application shell
+                  parentNavigatorKey: _rootNavigatorKey,
+                  builder: (BuildContext context, GoRouterState state) {
+                    return const SeedScreen();
+                  },
+                ),
+                GoRoute(
+                    path: CreateInvoiceScreen.subRouteName,
+                    // Use root navigator so the screen overlays the application shell
+                    parentNavigatorKey: _rootNavigatorKey,
+                    builder: (BuildContext context, GoRouterState state) {
+                      return const CreateInvoiceScreen();
+                    },
+                    routes: [
+                      GoRoute(
+                        path: ShareInvoiceScreen.subRouteName,
+                        // Use root navigator so the screen overlays the application shell
+                        parentNavigatorKey: _rootNavigatorKey,
+                        builder: (BuildContext context, GoRouterState state) {
+                          return ShareInvoiceScreen(invoice: state.extra as ShareInvoice);
+                        },
+                      ),
+                    ]),
+                GoRoute(
+                  path: ScannerScreen.subRouteName,
+                  parentNavigatorKey: _rootNavigatorKey,
+                  builder: (BuildContext context, GoRouterState state) {
+                    return const ScannerScreen();
+                  },
+                ),
+                GoRoute(
+                    path: WalletSettingsScreen.subRouteName,
+                    parentNavigatorKey: _rootNavigatorKey,
+                    builder: (BuildContext context, GoRouterState state) {
+                      return const WalletSettingsScreen();
+                    })
+              ],
+            ),
+            GoRoute(
+              path: TradeScreen.route,
+              builder: (BuildContext context, GoRouterState state) {
+                return const TradeScreen();
+              },
+              routes: <RouteBase>[
+                GoRoute(
+                    path: TradeSettingsScreen.subRouteName,
+                    parentNavigatorKey: _rootNavigatorKey,
+                    builder: (BuildContext context, GoRouterState state) {
+                      return const TradeSettingsScreen();
+                    })
+              ],
+            ),
+          ],
+        ),
+        GoRoute(
+            path: WelcomeScreen.route,
+            parentNavigatorKey: _rootNavigatorKey,
+            builder: (BuildContext context, GoRouterState state) {
+              return const WelcomeScreen();
+            },
+            routes: const []),
+      ],
+      redirect: (BuildContext context, GoRouterState state) async {
+        // TODO: It's not optimal that we read this from shared prefs every time, should probably be set through a provider
+        final hasEmailAddress = await Preferences.instance.hasEmailAddress();
+        if (!hasEmailAddress) {
+          FLog.info(text: "adding the email...");
+          return WelcomeScreen.route;
+        }
+
+        return null;
+      });
+
+  @override
+  void initState() {
+    super.initState();
+
+    init(context.read<bridge.Config>());
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    MaterialColor swatch = tenTenOnePurple;
+
+    return MaterialApp.router(
+      title: "10101",
+      theme: ThemeData(
+        primarySwatch: swatch,
+        iconTheme: IconThemeData(
+          color: tenTenOnePurple.shade800,
+          size: 32,
+        ),
+        extensions: <ThemeExtension<dynamic>>[
+          const TradeTheme(),
+          WalletTheme(colors: ColorScheme.fromSwatch(primarySwatch: swatch)),
+        ],
+      ),
+      routerConfig: _router,
+      debugShowCheckedModeBanner: false,
+    );
+  }
+
+  Future<void> init(bridge.Config config) async {
+    final orderChangeNotifier = context.read<OrderChangeNotifier>();
+    final positionChangeNotifier = context.read<PositionChangeNotifier>();
+    final candlestickChangeNotifier = context.read<CandlestickChangeNotifier>();
+    final walletChangeNotifier = context.read<WalletChangeNotifier>();
+    final channelService = context.read<ChannelInfoService>();
+    final initService = context.read<InitService>();
+
+    try {
+      initService.setupRustLogging();
+
+      subscribeToNotifiers(context);
+
+      initService.runBackend(config).then((_) {
+        FLog.info(text: "Backend started");
+      });
+
+      await orderChangeNotifier.initialize();
+      await positionChangeNotifier.initialize();
+      await candlestickChangeNotifier.initialize();
+
+      await logAppSettings(channelService, config);
+
+      final lastLogin = await initService.updateLastLogin();
+      FLog.debug(text: "Last login was at ${lastLogin.date}");
+    } on FfiException catch (error) {
+      FLog.error(text: "Failed to initialise: Error: ${error.message}", exception: error);
+    } catch (error) {
+      FLog.error(text: "Failed to initialise: $error", exception: error);
+    } finally {
+      FlutterNativeSplash.remove();
+    }
+    await walletChangeNotifier.refreshWalletInfo();
+  }
+}
+
+/// Forward the events from change notifiers to the Event service
+void subscribeToNotifiers(BuildContext context) {
+  // TODO: Move this code into an "InitService" or similar; we should not have bridge code in the widget
+
+  final EventService eventService = EventService.create();
+
+  final orderChangeNotifier = context.read<OrderChangeNotifier>();
+  final positionChangeNotifier = context.read<PositionChangeNotifier>();
+  final walletChangeNotifier = context.read<WalletChangeNotifier>();
+  final tradeValuesChangeNotifier = context.read<TradeValuesChangeNotifier>();
+  final submitOrderChangeNotifier = context.read<SubmitOrderChangeNotifier>();
+
+  eventService.subscribe(
+      orderChangeNotifier, bridge.Event.orderUpdateNotification(Order.apiDummy()));
+
+  eventService.subscribe(
+      submitOrderChangeNotifier, bridge.Event.orderUpdateNotification(Order.apiDummy()));
+
+  eventService.subscribe(
+      positionChangeNotifier, bridge.Event.positionUpdateNotification(Position.apiDummy()));
+
+  eventService.subscribe(
+      positionChangeNotifier,
+      const bridge.Event.positionClosedNotification(
+          bridge.PositionClosed(contractSymbol: bridge.ContractSymbol.BtcUsd)));
+
+  eventService.subscribe(
+      walletChangeNotifier, bridge.Event.walletInfoUpdateNotification(WalletInfo.apiDummy()));
+
+  eventService.subscribe(
+      tradeValuesChangeNotifier, bridge.Event.priceUpdateNotification(Price.apiDummy()));
+
+  eventService.subscribe(
+      positionChangeNotifier, bridge.Event.priceUpdateNotification(Price.apiDummy()));
+
+  eventService.subscribe(
+      AnonSubscriber((event) => FLog.info(text: event.field0)), const bridge.Event.log(""));
+}
+
+Future<void> logAppSettings(ChannelInfoService channelService, bridge.Config config) async {
+  String commit = const String.fromEnvironment('COMMIT');
+  if (commit.isNotEmpty) {
+    FLog.info(text: "Built on commit: $commit");
+  }
+
+  String branch = const String.fromEnvironment('BRANCH');
+  if (branch.isNotEmpty) {
+    FLog.info(text: "Built on branch: $branch");
+  }
+
+  PackageInfo packageInfo = await PackageInfo.fromPlatform();
+  FLog.info(text: "Build number: ${packageInfo.buildNumber}");
+  FLog.info(text: "Build version: ${packageInfo.version}");
+
+  FLog.info(text: "Network: ${config.network}");
+  FLog.info(text: "Esplora endpoint: ${config.esploraEndpoint}");
+  FLog.info(text: "Coordinator: ${config.coordinatorPubkey}@${config.host}:${config.p2PPort}");
+  FLog.info(text: "Oracle endpoint: ${config.oracleEndpoint}");
+  FLog.info(text: "Oracle PK: ${config.oraclePubkey}");
+
+  String nodeId = channelService.getNodeId();
+  FLog.info(text: "Node ID: $nodeId");
+}
